backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
paypalrestsdk==1.13.1
yfinance
beautifulsoup4
requests
pandas
numpy
google-generativeai
python-dotenv
pymysql==1.1.0
cryptography==41.0.7
